+++ llvm/include/llvm/IR/PassManagerInternal.h
@@ -19,6 +19,7 @@
 
 #include "llvm/ADT/STLExtras.h"
 #include "llvm/ADT/StringRef.h"
+#include "llvm/IR/Analysis.h"
 #include "llvm/Support/raw_ostream.h"
 #include <memory>
 #include <utility>
@@ -65,8 +66,8 @@ struct PassConcept {
 /// Can be instantiated for any object which provides a \c run method accepting
 /// an \c IRUnitT& and an \c AnalysisManager<IRUnit>&. It requires the pass to
 /// be a copyable object.
-template <typename IRUnitT, typename PassT, typename PreservedAnalysesT,
-          typename AnalysisManagerT, typename... ExtraArgTs>
+template <typename IRUnitT, typename PassT, typename AnalysisManagerT,
+          typename... ExtraArgTs>
 struct PassModel : PassConcept<IRUnitT, AnalysisManagerT, ExtraArgTs...> {
   explicit PassModel(PassT Pass) : Pass(std::move(Pass)) {}
   // We have to explicitly define all the special member functions because MSVC
@@ -84,8 +85,8 @@ struct PassModel : PassConcept<IRUnitT, AnalysisManagerT, ExtraArgTs...> {
     return *this;
   }
 
-  PreservedAnalysesT run(IRUnitT &IR, AnalysisManagerT &AM,
-                         ExtraArgTs... ExtraArgs) override {
+  PreservedAnalyses run(IRUnitT &IR, AnalysisManagerT &AM,
+                        ExtraArgTs... ExtraArgs) override {
     return Pass.run(IR, AM, ExtraArgs...);
   }
 
@@ -120,7 +121,7 @@ struct PassModel : PassConcept<IRUnitT, AnalysisManagerT, ExtraArgTs...> {
 ///
 /// This concept is parameterized over the IR unit that this result pertains
 /// to.
-template <typename IRUnitT, typename PreservedAnalysesT, typename InvalidatorT>
+template <typename IRUnitT, typename InvalidatorT>
 struct AnalysisResultConcept {
   virtual ~AnalysisResultConcept() = default;
 
@@ -139,7 +140,7 @@ struct AnalysisResultConcept {
   /// them. See the documentation in the \c AnalysisManager for more details.
   ///
   /// \returns true if the result is indeed invalid (the default).
-  virtual bool invalidate(IRUnitT &IR, const PreservedAnalysesT &PA,
+  virtual bool invalidate(IRUnitT &IR, const PreservedAnalyses &PA,
                           InvalidatorT &Inv) = 0;
 };
 
@@ -186,7 +187,7 @@ public:
 /// an invalidation handler. It is only selected when the invalidation handler
 /// is not part of the ResultT's interface.
 template <typename IRUnitT, typename PassT, typename ResultT,
-          typename PreservedAnalysesT, typename InvalidatorT,
+          typename InvalidatorT,
           bool HasInvalidateHandler =
               ResultHasInvalidateMethod<IRUnitT, ResultT>::Value>
 struct AnalysisResultModel;
@@ -194,10 +195,9 @@ struct AnalysisResultModel;
 /// Specialization of \c AnalysisResultModel which provides the default
 /// invalidate functionality.
 template <typename IRUnitT, typename PassT, typename ResultT,
-          typename PreservedAnalysesT, typename InvalidatorT>
-struct AnalysisResultModel<IRUnitT, PassT, ResultT, PreservedAnalysesT,
-                           InvalidatorT, false>
-    : AnalysisResultConcept<IRUnitT, PreservedAnalysesT, InvalidatorT> {
+          typename InvalidatorT>
+struct AnalysisResultModel<IRUnitT, PassT, ResultT, InvalidatorT, false>
+    : AnalysisResultConcept<IRUnitT, InvalidatorT> {
   explicit AnalysisResultModel(ResultT Result) : Result(std::move(Result)) {}
   // We have to explicitly define all the special member functions because MSVC
   // refuses to generate them.
@@ -220,7 +220,7 @@ struct AnalysisResultModel<IRUnitT, PassT, ResultT, PreservedAnalysesT,
   // FIXME: We should actually use two different concepts for analysis results
   // rather than two different models, and avoid the indirect function call for
   // ones that use the trivial behavior.
-  bool invalidate(IRUnitT &, const PreservedAnalysesT &PA,
+  bool invalidate(IRUnitT &, const PreservedAnalyses &PA,
                   InvalidatorT &) override {
     auto PAC = PA.template getChecker<PassT>();
     return !PAC.preserved() &&
@@ -233,10 +233,9 @@ struct AnalysisResultModel<IRUnitT, PassT, ResultT, PreservedAnalysesT,
 /// Specialization of \c AnalysisResultModel which delegates invalidate
 /// handling to \c ResultT.
 template <typename IRUnitT, typename PassT, typename ResultT,
-          typename PreservedAnalysesT, typename InvalidatorT>
-struct AnalysisResultModel<IRUnitT, PassT, ResultT, PreservedAnalysesT,
-                           InvalidatorT, true>
-    : AnalysisResultConcept<IRUnitT, PreservedAnalysesT, InvalidatorT> {
+          typename InvalidatorT>
+struct AnalysisResultModel<IRUnitT, PassT, ResultT, InvalidatorT, true>
+    : AnalysisResultConcept<IRUnitT, InvalidatorT> {
   explicit AnalysisResultModel(ResultT Result) : Result(std::move(Result)) {}
   // We have to explicitly define all the special member functions because MSVC
   // refuses to generate them.
@@ -255,7 +254,7 @@ struct AnalysisResultModel<IRUnitT, PassT, ResultT, PreservedAnalysesT,
   }
 
   /// The model delegates to the \c ResultT method.
-  bool invalidate(IRUnitT &IR, const PreservedAnalysesT &PA,
+  bool invalidate(IRUnitT &IR, const PreservedAnalyses &PA,
                   InvalidatorT &Inv) override {
     return Result.invalidate(IR, PA, Inv);
   }
@@ -267,16 +266,14 @@ struct AnalysisResultModel<IRUnitT, PassT, ResultT, PreservedAnalysesT,
 ///
 /// This concept is parameterized over the IR unit that it can run over and
 /// produce an analysis result.
-template <typename IRUnitT, typename PreservedAnalysesT, typename InvalidatorT,
-          typename... ExtraArgTs>
+template <typename IRUnitT, typename InvalidatorT, typename... ExtraArgTs>
 struct AnalysisPassConcept {
   virtual ~AnalysisPassConcept() = default;
 
   /// Method to run this analysis over a unit of IR.
   /// \returns A unique_ptr to the analysis result object to be queried by
   /// users.
-  virtual std::unique_ptr<
-      AnalysisResultConcept<IRUnitT, PreservedAnalysesT, InvalidatorT>>
+  virtual std::unique_ptr<AnalysisResultConcept<IRUnitT, InvalidatorT>>
   run(IRUnitT &IR, AnalysisManager<IRUnitT, ExtraArgTs...> &AM,
       ExtraArgTs... ExtraArgs) = 0;
 
@@ -289,10 +286,10 @@ struct AnalysisPassConcept {
 /// Can wrap any type which implements a suitable \c run method. The method
 /// must accept an \c IRUnitT& and an \c AnalysisManager<IRUnitT>& as arguments
 /// and produce an object which can be wrapped in a \c AnalysisResultModel.
-template <typename IRUnitT, typename PassT, typename PreservedAnalysesT,
-          typename InvalidatorT, typename... ExtraArgTs>
-struct AnalysisPassModel : AnalysisPassConcept<IRUnitT, PreservedAnalysesT,
-                                               InvalidatorT, ExtraArgTs...> {
+template <typename IRUnitT, typename PassT, typename InvalidatorT,
+          typename... ExtraArgTs>
+struct AnalysisPassModel
+    : AnalysisPassConcept<IRUnitT, InvalidatorT, ExtraArgTs...> {
   explicit AnalysisPassModel(PassT Pass) : Pass(std::move(Pass)) {}
   // We have to explicitly define all the special member functions because MSVC
   // refuses to generate them.
@@ -311,14 +308,12 @@ struct AnalysisPassModel : AnalysisPassConcept<IRUnitT, PreservedAnalysesT,
 
   // FIXME: Replace PassT::Result with type traits when we use C++11.
   using ResultModelT =
-      AnalysisResultModel<IRUnitT, PassT, typename PassT::Result,
-                          PreservedAnalysesT, InvalidatorT>;
+      AnalysisResultModel<IRUnitT, PassT, typename PassT::Result, InvalidatorT>;
 
   /// The model delegates to the \c PassT::run method.
   ///
   /// The return is wrapped in an \c AnalysisResultModel.
-  std::unique_ptr<
-      AnalysisResultConcept<IRUnitT, PreservedAnalysesT, InvalidatorT>>
+  std::unique_ptr<AnalysisResultConcept<IRUnitT, InvalidatorT>>
   run(IRUnitT &IR, AnalysisManager<IRUnitT, ExtraArgTs...> &AM,
       ExtraArgTs... ExtraArgs) override {
     return std::make_unique<ResultModelT>(
